fix minor compiler warnings and a sizeof doesn't necessairly return int
Sat Jan 15 00:53:48 2000  George Lebl <jirka@5z.com>

	* gnome-dentry.c, gnome-exec.c, gnome-moz-remote2.c, gnome-util.c:
	  fix minor compiler warnings and a sizeof doesn't necessairly
	  return int error.

+++ libgnome/gnome-exec.c
@@ -175,7 +175,7 @@ gnome_execute_async_with_env_fds (const char *dir, int argc,
   res = read (parent_comm_pipes[0], &child_pid, sizeof(child_pid));
   if (res != sizeof(child_pid))
     {
-      g_message("res is %d instead of %d", res, sizeof(child_pid));
+      g_message("res is %d instead of %d", res, (int)sizeof(child_pid));
       child_pid = -1; /* really weird things happened */
     }
   else if (read (parent_comm_pipes[0], &child_errno, sizeof(child_errno))
